Added graveyards
The game board was extended by 4 additional rows, which are displayed on the respective team's side. When a piece is killed, it is no longer disabled but is instead resized and moved to a graveyard.

Currently, this breaks the execution layer when a new game is started from an open instance of a game (I think because the board matrix is too large for older code).