Complete Implementation of Dice Roll (not final art)
Went ahead and completed the implementation of the dice roll. This means that any numbered rolls are supported and will display something, in case the animation is gone ahead and hooked up to functions.

The art, however, is not final. Animations for 1-4 are complete, 5 is only far enough along to differentiate it from 6. All animations will pop up with the correct face at the end of the animation, however.