Various Patches
AI now has a check in its UI movement so it does not capture its own pieces. (This itself could result in other bugs, if an AI piece manages to end its destination on a cell that already contains an AI piece)

Started work on staggered the dice roll animation between gameplay, not yet implemented.